chore: move types and service to corresponding places
