+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/components/mobile/available-jobs-assign-job-button-mobile.tsx
@@ -2,7 +2,7 @@ import { t } from 'i18next';
 import { TableButton } from '@/shared/components/ui/table-button';
 import { useJobsNotifications } from '../../../hooks';
 import { useAssignJobMutation } from '../../hooks/use-assign-job';
-import { type AssignJobBody } from '../../../services/jobs.service';
+import { type AssignJobBody } from '../../../types';
 
 export function AvailableJobsAssignJobButtonMobile({
   assignJobPayload,
