+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/hooks/use-refresh-jobs.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { jobsService, type RefreshJobsBody } from '../../services/jobs.service';
+import { jobsService } from '../../services/jobs.service';
+import { type RefreshJobsBody } from '../../types';
 
 export const refreshTasksSchema = z.unknown();
