+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/hooks/use-resign-job.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { jobsService, type RejectTaskBody } from '../../services/jobs.service';
+import { jobsService } from '../../services/jobs.service';
+import { type RejectTaskBody } from '../../types';
 
 export const rejectTaskSchema = z.unknown();
